# swagger.yaml
openapi: 3.0.0
info:
  title: invaPlay Backend API
  version: 1.0.0
  description: API documentation for the invaPlay Backend application, managing users, students, schools, classes, and grades.
servers:
  - url: http://localhost:5000 # Assuming the API is served from the root
  - url: https://api.innovplay.online/
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /skills:
    get:
      summary: Get list of all skills
      responses:
        '200':
          description: A list of all skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/schemas/Skill'
        '500':
          $ref: '#/responses/InternalServerError'
...
  /questions:
    get:
      summary: Get list of questions
      parameters:
        - name: test_id
          in: query
          schema:
            type: string
          description: Filter questions by test ID
        - name: type
          in: query
          schema:
            type: string
          description: Filter questions by type
        - name: subject_id
          in: query
          schema:
            type: string
          description: Filter questions by subject ID
        - name: created_by
          in: query
          schema:
            type: string
          description: Filter questions by creator
      responses:
        '200':
          description: A list of questions or a complete test with questions
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/schemas/Question'
                  - $ref: '#/schemas/TestWithQuestions'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '404':
          $ref: '#/responses/NotFoundError'
        '500':
          $ref: '#/responses/InternalServerError'
    post:
      summary: Create a new question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                type:
                  type: string
                subjectId:
                  type: string
                grade:
                  type: string
                createdBy:
                  type: string
                formattedText:
                  type: string
                secondStatement:
                  type: string
                title:
                  type: string
                description:
                  type: string
                command:
                  type: string
                subtitle:
                  type: string
                options:
                  type: array
                  items:
                    type: object
                skills:
                  type: array
                difficulty:
                  type: string
                solution:
                  type: string
                formattedSolution:
                  type: string
                test_id:
                  type: string
                value:
                  type: number
                topics:
                  type: array
                version:
                  type: integer
                lastModifiedBy:
                  type: string
                educationStageId:
                  type: string
              required:
                - text
                - type
                - subjectId
                - grade
                - createdBy
      responses:
        '201':
          description: Question created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
        '400':
          $ref: '#/responses/BadRequestError'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '500':
          $ref: '#/responses/InternalServerError'
    delete:
      summary: Bulk delete questions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question_ids:
                  type: array
                  items:
                    type: string
              required:
                - question_ids
      responses:
        '200':
          description: Questions deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/responses/BadRequestError'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '500':
          $ref: '#/responses/InternalServerError'
  /questions/{question_id}:
    get:
      summary: Get question details by ID
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: string
          description: ID of the question
      responses:
        '200':
          description: Question details
          content:
            application/json:
              schema:
                $ref: '#/schemas/Question'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '404':
          $ref: '#/responses/NotFoundError'
        '500':
          $ref: '#/responses/InternalServerError'
    put:
      summary: Update question by ID
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: string
          description: ID of the question to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                type:
                  type: string
                subjectId:
                  type: string
                grade:
                  type: string
                formattedText:
                  type: string
                secondStatement:
                  type: string
                title:
                  type: string
                description:
                  type: string
                command:
                  type: string
                subtitle:
                  type: string
                options:
                  type: array
                  items:
                    type: object
                skills:
                  type: array
                difficulty:
                  type: string
                solution:
                  type: string
                formattedSolution:
                  type: string
                value:
                  type: number
                topics:
                  type: array
                version:
                  type: integer
                lastModifiedBy:
                  type: string
                educationStageId:
                  type: string
      responses:
        '200':
          description: Question updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/responses/BadRequestError'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '404':
          $ref: '#/responses/NotFoundError'
        '500':
          $ref: '#/responses/InternalServerError'
    delete:
      summary: Delete a question by ID
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: string
          description: ID of the question to delete
      responses:
        '200':
          description: Question deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '404':
          $ref: '#/responses/NotFoundError'
        '500':
          $ref: '#/responses/InternalServerError'
...
responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Mensagem explicando o motivo da falha de autenticação.
                example: Token inválido ou expirado.
    ForbiddenError:
      description: User does not have the necessary permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Mensagem indicando que o usuário não tem permissão.
                example: Acesso negado para o recurso solicitado.
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensagem sobre o recurso não encontrado.
                example: Recurso solicitado não foi encontrado.
    BadRequestError:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensagem descrevendo o erro de validação.
                example: Dados inválidos enviados na requisição.
              details:
                type: object # Can be a string or a more detailed object
                description: Detalhes adicionais sobre o erro.
                example:
                  campo: email
                  problema: Formato inválido.
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensagem genérica de erro.
                example: Ocorreu um erro inesperado.
              details:
                type: string
                description: Descrição detalhada do erro para depuração.
                example: Consulte os logs do servidor para mais informações.